23/06/05 추가

+++ src/main/java/kit/item/dto/entity/market/SaleProductReviewInfoDto.java
@@ -19,5 +19,5 @@ public class SaleProductReviewInfoDto {
     String ownerName;
     LocalDateTime date;
     String comment;
-    Double rating;
+    Long rating;
 }
